+++ src/main/java/com/sq_yan/magic_storage/registry/MSBlocks.java
@@ -1,7 +1,8 @@
 package com.sq_yan.magic_storage.registry;
 
 import com.sq_yan.magic_storage.MagicStorage;
-import com.sq_yan.magic_storage.block.HeartStorageBlock;
+import com.sq_yan.magic_storage.block.HeartStorageT1Block;
+import com.sq_yan.magic_storage.block.HeartStorageT2Block;
 import com.sq_yan.magic_storage.block.StorageCellBlock;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.SoundType;
@@ -14,11 +15,21 @@ import net.minecraftforge.registries.RegistryObject;
 public final class MSBlocks {
     public static final DeferredRegister<Block> REGISTRY = DeferredRegister.create(ForgeRegistries.BLOCKS, MagicStorage.MODID);
 
-    public static final RegistryObject<HeartStorageBlock> HEART_STORAGE = REGISTRY.register("heart_storage",
-        () -> new HeartStorageBlock(BlockBehaviour.Properties.of()
+    public static final RegistryObject<HeartStorageT1Block> HEART_STORAGE = REGISTRY.register("heart_storage",
+        () -> new HeartStorageT1Block(BlockBehaviour.Properties.of()
             .setId(REGISTRY.key("heart_storage"))
             .mapColor(MapColor.COLOR_PURPLE)
             .strength(5.0f, 6.0f)
+            .sound(SoundType.METAL)
+            .requiresCorrectToolForDrops()
+        )
+    );
+
+    public static final RegistryObject<HeartStorageT2Block> HEART_STORAGE_T2 = REGISTRY.register("heart_storage_t2",
+        () -> new HeartStorageT2Block(BlockBehaviour.Properties.of()
+            .setId(REGISTRY.key("heart_storage_t2"))
+            .mapColor(MapColor.COLOR_PURPLE)
+            .strength(6.0f, 8.0f)
             .sound(SoundType.METAL)
             .requiresCorrectToolForDrops()
         )